Models associations
Was completed the associations between each model. But also was remove
the image foreign keys and indexes for the images_downloads and
images_likes, due to only a user can access to her images that were
download or liked for others users.
It is easiest to create only a FK for the user in images_downloads and
images_likes and then de user can see her theri images posted, their
images downloaded by others users or downloades by himself, and their
images liked by others uses or liked by himself.