Image post update
The form.html.erb file was changed into the iamges folder, this file
will be use for upload a new image and for update an image.
The edit view renders this form and shows the current image for that id,
the 'new' template also renders this form.

In the image controller was change the edit method, was deleted the
content because was setted up into to the top of the class set an image,
this calls the method 'set_image' and find the image with id as param.

In the routes file was remove the only specifications, instead was added
and except specification, this create all the necessary resources
excepting for show view.