Image like functionality
Was created a the likes controller, this controller has the like method,
that gets the image_liked with the id that comes as parameter.

After that, it validates if the current user already liked that image,
if true, then unlikes it, otherwise likes the image.

In the index view was added a div with a button_to, that basically is a
form to the '/like/:id', with post method.